Добавил абстракцию
