+++ src/view/edit-event-item-view.js
@@ -2,7 +2,7 @@
 import dayjs from 'dayjs';
 import {locations} from '../mock/locations';
 import {eventTypes} from '../mock/event-types';
-import { createElement } from '../render';
+import AbstractView from './abstract-view';
 
 const createEventItemEditTemplate = (tripEvent) => {
   const {eventType, cost, location, startDate, endDate, offers, description} = tripEvent;
@@ -102,27 +102,15 @@ const createEventItemEditTemplate = (tripEvent) => {
             </li>`;
 };
 
-export default class EventItemEditView {
-  #element = null;
+export default class EventItemEditView extends AbstractView {
   #event = null;
 
   constructor(event) {
+    super();
     this.#event = event;
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
   }
 
   get template() {
     return createEventItemEditTemplate(this.#event);
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
