+++ src/view/event-list-view.js
@@ -1,23 +1,10 @@
-import { createElement } from '../render';
+import AbstractView from './abstract-view';
+
 
 const createEventsListTemplate = () => ('<ul class="trip-events__list"></ul>');
 
-export default class EventsListView {
-  #element = null;
-
-  get element() {
-    if(!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class EventsListView extends AbstractView {
   get template() {
     return createEventsListTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
