+++ src/view/trip-filters-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render';
+import AbstractView from './abstract-view';
 
 const createTripFiltersTemplate = () => (
   `<form class="trip-filters" action="#" method="get">
@@ -18,23 +18,8 @@ const createTripFiltersTemplate = () => (
               </form>`
 );
 
-
-export default class TripFiltersView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class FilterView extends AbstractView {
   get template() {
-    return createTripFiltersTemplate();
-  }
-
-  removeElementO() {
-    this.#element = null;
+    return createTripFiltersTemplate;
   }
 }
